pyproject.toml
[project]
name = "docling_graph"
version = "0.1.0"
description = "A tool to convert documents into knowledge graphs using Docling."

license = { text = "MIT License" }
keywords = ["docling", "knowledge-graph", "nlp", "pdf", "graph"]
classifiers = [
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",    
]
readme = "README.md"
authors = [ # In alphabetical order
  { name = "Ayoub El Bouchtili", email = "ayoub.elbouchtili@fr.ibm.com" },
  { name = "Michele Dolfi", email = "dol@zurich.ibm.com" },
  { name = "Maxime Gillot", email = "Maxime.Gillot@ibm.com" },  
  { name = "Sophie Lang", email = "sophie.lang@de.ibm.com" },
  { name = "Guilhaume Leroy Meline", email = "guilhaume@fr.ibm.com" },
  { name = "Peter Staar", email = "taa@zurich.ibm.com" },
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "docling-core[chunking]>=2.50.0,<3.0.0",
    "docling[vlm]>=2.0.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "networkx>=3.0.0,<4.0.0",
    "rich>=13,<14",  # verify constraints you need
    "typer[all]>=0.12,<1.0.0",  # verify min version used in code
    "python-dotenv>=1.0,<2.0",
]

[project.urls]
homepage = "https://github.com/ayoub-ibm/docling-graph"
repository = "https://github.com/ayoub-ibm/docling-graph"
issues = "https://github.com/ayoub-ibm/docling-graph/issues"
changelog = "https://github.com/ayoub-ibm/docling-graph/blob/main/CHANGELOG.md"

[project.scripts]
docling-graph = "docling_graph.cli.main:main"

# LLM-related optional dependencies
llm = "docling_graph.llm:main"

[dependency-groups]
# Development tools (linters, type-checkers) and utilities
dev = [
    "pre-commit~=3.7",
    "mypy~=1.10",
    "types-setuptools~=70.3",
    "pandas-stubs~=2.1",
    "types-openpyxl~=3.1",
    "types-requests~=2.31",
    "boto3-stubs~=1.37",
    "types-urllib3~=1.26",
    "types-tqdm~=4.67",
    "coverage~=7.6",
    "pytest~=8.3",
    "pytest-cov>=6.1.1",
    "pytest-dependency~=0.6",
    "pytest-durations~=1.6.1",
    "pytest-xdist~=3.3",
    "ipykernel~=6.29",
    "nbqa~=1.9",
    "python-semantic-release~=7.32",
    "mypy>=1.4.0",
    "ruff>=0.0.280",
]

# Test stack and utilities
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-xdist>=3.3.0",
    "pytest-timeout>=2.1.0",
    "hypothesis>=6.82.0",
    "pytest-asyncio>=0.21.0",
    "freezegun>=1.2.2",
    "responses>=0.23.0",
    "faker>=19.0.0",
    "types-PyYAML",
    "types-requests",
    "mypy>=1.4.0",
]

# Notebook-related tools and data analysis helpers
notebook = [
    "notebook",
    "ipython",
    "pandas",
]

[tool.uv]
package = true
default-groups = "all"

[tool.setuptools.packages.find]
include = ["docling_graph*"]

[tool.ruff]
target-version = "py310"
line-length = 100
respect-gitignore = true

[tool.ruff.format]
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "W",    # pycodestyle warnings
  "F",    # pyflakes
  "I",    # isort
  "N",    # pep8-naming
  "UP",   # pyupgrade
  "B",    # flake8-bugbear
  "A",    # flake8-builtins
  "C4",   # flake8-comprehensions
  "C90",  # mccabe (complexity)
  "ANN",  # flake8-annotations
  "PD",   # pandas-vet
  "PIE",  # flake8-pie
  "Q",    # flake8-quotes
  "RUF",  # Ruff-specific checks
  "ASYNC" # async
]
ignore = [
  "C408",   # Unnecessary `dict()` call (rewrite as a literal)
  "E501",   # Line too long, handled by ruff formatter
  "D107",   # "Missing docstring in __init__",
  "F401",   # imported but unused; consider using `importlib.util.find_spec` to test for "
  "F811",   # "redefinition of the same function"
  "PL",     # Pylint
  "RUF012", # Mutable Class Attributes
  "UP006",  # List vs list, etc
  "UP007",  # Option and Union
  "UP035",  # `typing.Set` is deprecated, use `set` instead"
  # "ANN101",  # Missing type annotation for self
  # "ANN102",  # Missing type annotation for cls
  "ANN401",  # Dynamically typed expressions (Any)
]

[tool.ruff.lint.pep8-naming]
classmethod-decorators = [
  # Allow Pydantic's `@validator` decorator to trigger class method treatment.
  "pydantic.validator",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"tests/*.py" = ["ASYNC"]

[tool.ruff.lint.mccabe]
max-complexity = 30

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.mypy]
pretty = true
python_version = "3.10"
no_implicit_optional = true
plugins = "pydantic.mypy"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "docling.*",
    "rich.*",
    "typer.*",
    "yaml.*",
    "mypy.*",
    "pandas.*",
    "networkx.*",
    "ollama.*",
    "openai.*",
    "mistralai.*",
    "google-genai.*",
    "google.*"
]
ignore_missing_imports = true

[tool.semantic_release]
# for default values check:
# https://github.com/python-semantic-release/python-semantic-release/blob/v7.32.2/semantic_release/defaults.cfg

# version_source = "tag_only"
# branch = "main"

# configure types which should trigger minor and patch version bumps respectively
# (note that they must be a subset of the configured allowed types):
# parser_angular_allowed_types = "build,chore,ci,docs,feat,fix,perf,style,refactor,test"
# parser_angular_minor_types = "feat"
# parser_angular_patch_types = "fix,perf"